feat: landlock koboldcpp
Add landlock-koboldcpp shell script

Some features require additional permissions (i.e using a port that is
not 5001); Other features are not intuitively supported by landlock's
architecture (i.e "Launch Browser" with xdg-open).

It should be noted that INT , or SIGINT , is trapped. This is due to the
program not having a graceful way to exit after a server is started.
Without this trap, cleanup may not occur.

It should also be noted some non-standard directories are used (i.e
$HOME/.local/share/models). Remove these directories if they are not
needed.